pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
backend-path = ["."]

[project]
name = "tanjun_5.0"
version = "1.1.1"
description = "The Officila Tanjun Discord Bot!"
readme = "README.md"
requires-python = ">=3.12"

license = {file = "LICENSE.txt"}

authors = [
    {name = "EntchenEric", email = "eseidel2004@gmail.com"},
    {name = "Arion2000", email = "github@arion2000.xyz"}
]

dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiomysql==0.2.0",
    "aiosignal==1.3.2",
    "akinator-python==1.6.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "asyncmy==0.2.10",
    "attrs==25.3.0",
    "backcall==0.2.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "colorama==0.4.6",
    "contourpy==1.3.2",
    "cryptography==44.0.3",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "Deprecated==1.2.18",
    "discord.py==2.5.2",
    "distro==1.9.0",
    "docopt==0.6.2",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "fonttools==4.57.0",
    "frozenlist==1.6.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "jedi==0.19.2",
    "Jinja2==3.1.6",
    "jiter==0.9.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter_client==8.6.3",
    "jupyter_core==5.7.2",
    "jupyterlab_pygments==0.3.0",
    "kiwisolver==1.4.8",
    "MarkupSafe==3.0.2",
    "matplotlib==3.10.1",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "mysql-connector-python==9.2.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "num2words==0.5.14",
    "numpy==2.2.5",
    "openai==1.69.0",
    "packaging==25.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pickleshare==0.7.5",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "prompt_toolkit==3.0.51",
    "propcache==0.3.1",
    "pure_eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.11.1",
    "pydantic_core==2.33.0",
    "PyGithub==2.6.1",
    "Pygments==2.19.1",
    "PyJWT==2.10.1",
    "PyMySQL==1.1.1",
    "PyNaCl==1.5.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyzmq==26.4.2",
    "referencing==0.36.2",
    "requests==2.32.3",
    "rpds-py==0.24.0",
    "scipy==1.15.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "tinycss2==1.4.0",
    "tornado==6.4.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "twitchAPI==4.4.0",
    "typing_extensions==4.13.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "webencodings==0.5.1",
    "wrapt==1.17.2",
    "yarg==0.1.10",
    "yarl==1.20.0",
]

[project.urls]
"Homepage" = "https://docs.tanjun.bot/"
"Repository" = "https://github.com/TanjunBot/new_tanjun"
"Bug Tracker" = "https://github.com/TanjunBot/new_tanjun/issues"

[tool.ruff]
target-version = "py312"

line-length = 127

# E = pycodestyle errors
# W = pycodestyle warnings
# F = Pyflakes
# I = isort
# C901 = McCabe complexity 
# UP = pyupgrade
# A = flake8-builtins
# B = flake8-bugbear
# SIM = flake8-simplify
# TID = flake8-tidy-imports
# N = pep8-naming
select = ["E", "W", "F", "I", "C901", "UP", "A", "B", "SIM", "TID", "N"]

# dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.format]
# Zeilenlänge für den Formatter

quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false
